Warning Fixes - const correctness in unicode encoding, unused variables in blenlib, and some type conversions
This is from a patch that is in the tracker, but it leaves out a fix of BLI_gzopen which needs more work.

+++ source/gameengine/Converter/BL_BlenderDataConversion.cpp
@@ -2363,8 +2363,8 @@ void BL_ConvertBlenderObjects(struct Main* maggie,
 			frame_type = RAS_FrameSettings::e_frame_scale;
 		}
 		
-		aspect_width = blenderscene->r.xsch*blenderscene->r.xasp;
-		aspect_height = blenderscene->r.ysch*blenderscene->r.yasp;
+		aspect_width  = (int)(blenderscene->r.xsch * blenderscene->r.xasp);
+		aspect_height = (int)(blenderscene->r.ysch * blenderscene->r.yasp);
 	}
 	
 	RAS_FrameSettings frame_settings(
